Moving all images to assets folder
