+++ docs/Components/SudokuBoard/HighlightColorsDiagram.png
